requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.8.0
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
flake8==7.2.0
iniconfig==2.1.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.1.0
pytest-django==4.11.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.0
tzdata==2025.2
